+++ neonpay/adapters/__init__.py
@@ -7,11 +7,13 @@ from .aiogram_adapter import AiogramAdapter
 from .ptb_adapter import PythonTelegramBotAdapter
 from .telebot_adapter import TelebotAdapter
 from .raw_api_adapter import RawAPIAdapter
+from .botapi_adapter import BotAPIAdapter
 
 __all__ = [
     "PyrogramAdapter",
     "AiogramAdapter", 
     "PythonTelegramBotAdapter",
     "TelebotAdapter",
-    "RawAPIAdapter"
+    "RawAPIAdapter",
+    "BotAPIAdapter",
 ]